Move the BarButtonItem class into the ReaderViewController file for now, for simplicity.
